Revert "Merge pull request #37 from shared-systems/martin/new_login"
This reverts commit e072abab62cc8826f05d8a887e762665a014ecd3, reversing
changes made to 2b67b583efb07fd7cc1e9a23198ddf9984ea3660.
