+++ format_standardization.txt
@@ -1,0 +1,65 @@
+***************************************************
+
+File format:
+
+    # file description
+
+    # library import
+    #   external library import
+    #   (1 empty line)
+    #   internal library import
+
+    # (library to class/function, 2 empty line)
+
+    class xxx(xxx):
+
+        self.param = xxx
+
+        def __init__(self, parent: Object=None, *args, **kwargs):
+            super(self, <Replace with parent class name>).__init__(*args, **kwargs)
+
+            # define all class object that will be use on class functions
+
+            # For example:
+            # self.var = None
+
+        # (function/class object to next one, 1 empty line)
+
+        def _func(param):
+            # function that used internally begin with "_"
+            pass
+
+        def func1(param):
+            self.var = 1
+            print(self.var)
+            pass
+
+        # (function/class object to next one, 1 empty line)
+
+        def func2(param):
+            pass
+
+    # (class/function to class/function, 2 empty line)
+
+    def func(param):
+        pass
+
+    # each function need to have description at top about what it helps on
+
+    # description
+    def helper_func(param):
+        pass
+
+    # (1 empty line at bottom)
+
+***************************************************
+
+PyQt module import:
+
+    from PyQt5.<Sub_Module_Name> import <Feature_Name>
+
+For example:
+
+    from PyQt5.QtWidgets import QMainWindow
+
+***************************************************
