+++ src/resources.py
@@ -32,14 +32,14 @@
         self.current_row = 0        # param number
 """
 
-from src.interfaces.interface import Interface
+from src.mainview import MainView
 from src.uix.util import *
 from src.uix.config import *
 from src.uix.popup import Question
 from src.api import Api
 
 
-class Resources(Interface):
+class Resources(MainView):
 
     def __init__(self, *args, **kwargs):
         super(Resources, self).__init__(*args, **kwargs)
